A Spring Boot-based e-commerce platform with JWT authentication, Iyzico payment integration, and Thymeleaf & JSON support for front-end and back-end communication.
This is a Spring Boot-based e-commerce platform with JWT authentication for secure user login and registration. It includes Iyzico payment integration for handling transactions and supports both Thymeleaf for rendering the front-end and JSON for back-end communication with the front-end. The application enables users to browse products, manage their cart, place orders, and handle payments in a secure and seamless environment.